requirements-parser.txt
torch>=1.7.1
pytorch-lightning>=1.1.1
pyyaml>=5.3.1
tensorboard
optuna>=2.3.0
transformers>=4.1.1
fasttext==0.9.1
gdown
